+++ lib/glimmer/dsl/web/observe_expression.rb
@@ -20,7 +20,7 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 require 'glimmer/dsl/static_expression'
-# require 'glimmer/dsl/top_level_expression'
+require 'glimmer/dsl/top_level_expression'
 require 'glimmer/dsl/observe_expression'
 require 'glimmer/web/component'
 
@@ -28,7 +28,7 @@ module Glimmer
   module DSL
     module SWT
       class ObserveExpression < StaticExpression
-#         include TopLevelExpression # TODO delete if not needed
+        include TopLevelExpression
         include Glimmer::DSL::ObserveExpression
 
         def interpret(parent, keyword, *args, &block)